requirements.txt
parse-nmap
python-libnmap
termcolor
